+++ src/mosaico_ip_agent/__init__.py
@@ -19,5 +19,5 @@
 #
 #  SPDX-License-Identifier: MIT
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 __author__ = "André S. Gomes"